requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
boto3==1.11.13
botocore==1.14.13
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-crispy-forms==1.8.1
django-storages==1.9.1
docutils==0.15.2
google==2.0.3
gunicorn==20.0.4
idna==2.8
jmespath==0.9.4
nltk==3.4.5
numpy==1.18.1
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.9
